Fixed compiler warnings in sample code.

+++ samples/dnn/layer_example_f32.c
@@ -90,7 +90,7 @@ LIBXSMM_INLINE void zero_buf(float* buf, size_t size) {
 #if defined(_OPENMP)
 # pragma omp parallel for private(i)
 #endif
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = 0.0f;
   }
 }
@@ -100,7 +100,7 @@ LIBXSMM_INLINE void copy_buf(float* src, float* dst, size_t size) {
 #if defined(_OPENMP)
 # pragma omp parallel for private(i)
 #endif
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     dst[i] = src[i];
   }
 }
@@ -109,7 +109,7 @@ LIBXSMM_INLINE void init_buf(float* buf, size_t size, int initPos, int initOne)
 {
   int i;
   zero_buf(buf, size);
-  for (i = 0; i < size; ++i) {
+  for (i = 0; i < (int)size; ++i) {
     buf[i] = (float)((initOne != 0) ? 1.0 : ((initPos != 0) ? drand48() : (0.05 - drand48()/10.0)));
   }
 }
